+++ openerp_proxy/ext/repr/utils.py
@@ -1,6 +1,8 @@
 import os
 import os.path
+import collections
 
+from jinja2 import Template
 
 from ...utils import ustr as _
 from ...utils import makedirs
@@ -21,58 +23,61 @@ makedirs(REPORTS_PATH)
 
 
 # HTML Templates
-TMPL_INFO_WITH_HELP = u"""
+TMPL_OBJECT_DESCRIPTION = Template("""
 <div class="container-fluid">
     <div class="row">
         <div class="panel panel-default col-md-7 col-lg-7">
-            <div class="panel-heading">%(caption)s</div>
-            <div class="panel-body">%(info)s</div>
+            <div class="panel-heading">{{ caption }}</div>
+            <div class="panel-body">
+                <table class="table table-bordered table-condensed {{ extra_classes }}"
+                       style="margin-left:0; {{ styles }}">
+                    {% if headers %}
+                        <tr>
+                            {% for header in headers %}
+                                <th>{{ header }}</th>
+                            {% endfor %}
+                        </tr>
+                    {% endif %}
+                    {% for row in data %}
+                    <tr>
+                        <th>{{ row[0] }}</th>
+                        {% for cell in row[1:] %}
+                            <td>{{ cell }}</td>
+                        {% endfor %}
+                    </tr>
+                    {% endfor %}
+                </table>
+            </div>
         </div>
         <div class="panel panel-default col-md-5 col-lg-5">
             <div class="panel-heading">Info</div>
-            <div class="panel-body">%(help)s</div>
+            <div class="panel-body">{{ help }}</div>
         </div>
     </div>
 </div>
-"""
-
-TMPL_TABLE = u"""
-<table class="table table-bordered table-condensed %(extra_classes)s" style="margin-left:0;%(styles)s">
-%(rows)s
-</table>
-"""
-
-TMPL_TABLE_ROW = u"<tr>%s</tr>"
-TMPL_TABLE_DATA = u"<td>%s</td>"
-TMPL_TABLE_HEADER = u"<th>%s</th>"
+""")
 
 
-def th(val):
-    return TMPL_TABLE_HEADER % _(val)
-
-
-def td(val):
-    return TMPL_TABLE_DATA % _(val)
-
-
-def tr(*args):
-    return TMPL_TABLE_ROW % u"".join(args)
-
-
-def describe_object_html(data, caption='', help='', table_styles=''):
+def describe_object_html(data, caption='', help='', table_styles='',
+                         headers=None):
     """ Converts dictionary data to html table string
 
-        :param dict data: dictionary like object. must contain .items() method
-                          represents object info to be displayed in table
+        :param dict|list data: dictionary like object.
+                               must contain .items() method
+                               represents object info to be displayed in table
+                               or iterable table like structure, where
+                               first element ofeach row is assumend
+                               to be header
         :param str caption: table's caption
         :param str help: help message to be displayed near table
         :param str table_styles: string with styles for table
     """
-    html_data = u"".join((tr(th(k), td(v)) for k, v in data.items()))
+    if isinstance(data, dict):
+        data = [(k, v) for k, v in data.items()]
 
-    table = TMPL_TABLE % {'styles': table_styles,
-                          'rows': html_data,
-                          'extra_classes': ''}
-    return TMPL_INFO_WITH_HELP % {'info': table, 'help': help, 'caption': caption}
-
-
+    return TMPL_OBJECT_DESCRIPTION.render(data=data,
+                                          styles=table_styles,
+                                          extra_classes='',
+                                          help=help,
+                                          caption=caption,
+                                          headers=headers)